Fixed incorrect integer type which can cause crashes
Changing the date viewed (earlier/negatively) by large amounts results
in a crash (caused by a stack overflow). This is caused by an unsigned
integer being used to hold a potentially negative number.

On the simulator, this doesn’t cause a problem as the unsigned integer
is then passed to a method that takes a sign integer. However, on a
real device this causes a crash.

+++ CalendarLib/MGCDayPlannerView.m
@@ -370,7 +370,7 @@ static const CGFloat kMaxHourSlotHeight = 150.;
 {
     CGFloat dayWidth = self.dayColumnSize.width;
 	
-	NSUInteger first = floorf(self.timedEventsView.contentOffset.x / dayWidth);
+	NSInteger first = (NSInteger)floorf(self.timedEventsView.contentOffset.x / dayWidth);
 	NSDate *firstDay = [self dateFromDayOffset:first];
 	if (self.dateRange && [firstDay compare:self.dateRange.start] == NSOrderedAscending)
 		firstDay = self.dateRange.start;
